src/p_and_r1/requirements.txt
black==23.9.1
click==8.1.7
exceptiongroup==1.1.3
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pytest==7.4.3
PyYAML==6.0.1
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.8.0
